fix linting issue

+++ crates/nu-cli/src/commands/math/sum.rs
@@ -77,7 +77,7 @@ fn to_byte(value: &Value) -> Option<Value> {
 fn to_duration(value: &Value) -> Option<Value> {
     match &value.value {
         UntaggedValue::Primitive(Primitive::Int(duration)) => Some(
-            UntaggedValue::Primitive(Primitive::Duration(BigInt::from((*duration).clone())))
+            UntaggedValue::Primitive(Primitive::Duration((*duration).clone()))
                 .into_untagged_value(),
         ),
         _ => None,
